Put non-default true/false parms back into the Galileo Gen1 board default settings.
Still not clear if these are needed, but they are clearly not the default values for those two parameters and so it appears they are required. No Gen1 handy to test this out, but it appears that the main goal was to use RAW mode to address the LED GPIO pin, rather than the header pin number. Not sure why this is the case, may be an historical artifact of this sample.