Need to try/catch require on mraa, even if it exists.
This is due to the fact that mraa includes many compiled parts. These parts are Node.js version-specific. If you’ve installed a version of mraa onto your system and then change the version of Node.js so that the entry points are no longer compatible you’ll encounter fatal errors when mraa is called. Thus, doing the require.resolve() check only lets you know that mraa is present, but does not help regarding this possible mismatch. The only way to find that is to do the require inside a try/catch.